requirements.txt
fastapi
uvicorn
psycopg2-binary
redis
langgraph
typing-extensions
google-generativeai

